Refactor: 기타 경로 수정 및 PostPage 버튼 수정
