--- openmind/src/Components/FloatingButton.jsx
@@ -1,11 +1,0 @@
-import Styles from '../Styles/FloatingButton.module.css';
-
-const FloatingButton = ({ children, ...rest }) => {
-  return (
-    <button className={Styles.floatingButton} {...rest}>
-      {children}
-    </button>
-  );
-};
-
-export default FloatingButton;
