+++ openmind/src/Components/UserCard.jsx
@@ -1,5 +1,5 @@
-import ICON_MESSAGE from '../../Assets/Icon/iconMessages.svg';
-import Styles from '../../Styles/UserCard.module.css';
+import ICON_MESSAGE from '../Assets/Icon/iconMessages.svg';
+import Styles from '../Styles/UserCard.module.css';
 import { useNavigate } from 'react-router-dom';
 
 const UserCard = ({ id, name, imageSource, questionCount }) => {
